[FIX] uptime data type change int16_t to int32_t and prevent uptime overflow

+++ include/format.h
@@ -5,7 +5,7 @@
 #include <string>
 
 namespace Format {
-std::string ElapsedTime(long times);  // TODO(@sangwon): See src/format.cpp
-};                                    // namespace Format
+std::string ElapsedTime(int32_t times);  // TODO(@sangwon): See src/format.cpp
+};                                       // namespace Format
 
 #endif
